Tue, Feb 28
We finally got the motor to go to a level and stay at that level (although we almost broke the motor in the process). Because of that, I've decided to merge mathew mode into our main mode (and also so Marcus doesn't kill me).

Mathew mode had the working arm code, and the normal repository had the working auton code.