Livonia Edits
This has edits that we made to the code at Livonia. This mainly consists of tweaking our middle auton and making last minute edits to the auto align function that Matthew never used, as well as minor edits to our LED code and having our drive motors start in coast rather than brake.