pre-saginaw edits
These were all the edits we made going into states. DriveDistanceInches and DriveDistanceRotations were merged into one file along with AutoTurnDegrees and AutoTurnRotations. We removed wheel movement for the autoAlign and created different pipelines for the top and middle poles. We also added the sendable chooser for auton, added an emergency stop for the arm, and added a ramp rate for the drive motors.

I also tried making a smoother autobalance.