Saginaw edits
This has all the edits we made at states. We TRIED making it so that the arm would only move backwards when the robot was moving (DefaultArmState.java).
We also made edits to our auton. I realized that the reason why I kept changing the autobalance multiplier was because it needed more speed going forwards than backwards, but I didn't have enough time to implement that.
We also fixed the AutoTurn command so it no longer resets the encoders.
NOTE: during the competition they removed a link from the arm chain, so all of our pot constants (except TopPositionCube) are messed up.